improved log readability

+++ R/updateRepo.R
@@ -11,48 +11,56 @@
 #' @param check Boolean deciding whether package must have been checked or not 
 #' in order to be distributed
 #' @param force_rebuild Option to rebuild all packages from source
+#' @param clean Option to clean repos before updating/pulling to avoid merge conflicts
 #' @author Jan Philipp Dietrich
 #' @seealso \code{\link{buildLibrary}}
 #' @export
 
-updateRepo <- function(path=".", check=TRUE, force_rebuild=FALSE) {
+updateRepo <- function(path=".", check=TRUE, force_rebuild=FALSE, clean=FALSE) {
   cat(date(),"\n")
   cwd <- getwd()
   on.exit(setwd(cwd))
   setwd(path)
   
   if(dir.exists(".svn")) {
-    system("svn revert -R .")
+    if(clean) system("svn revert -R .")
     system("svn update")
   }
   ap <- suppressWarnings(available.packages(paste0("file:",getwd())))
   
   dirs <- grep("^\\.",list.dirs(recursive = FALSE,full.names = FALSE), value=TRUE, invert=TRUE)
+  nchar <- max(length(dirs))
   for(d in dirs) {
+    fd <- format(d,width=nchar)
     curversion <- tryCatch(ap[d,"Version"],error=function(e)return(0))
     setwd(d)
     if(dir.exists(".svn")) {
-      system("svn revert -R .")
-      system("svn update")
+      if(clean) system("svn revert -Rq .")
+      system("svn update -q")
     }
     if(dir.exists(".git")) {
-      system("git reset --hard HEAD")
-      system("git --no-pager show -s --format='%an <%ae>' HEAD")
-      system("git clean -fxq")
-      system("git pull")
+      if(clean) {
+        system("git reset --hard HEAD -q")
+        system("git clean -fxq")
+      }
+      system("git pull -q")
     }
     vkey <- validkey()
     if(curversion < vkey$version | force_rebuild) {
       if(vkey$valid | !check | force_rebuild) {
         if(vkey$roxygen) suppressWarnings(devtools::document(pkg=".",roclets=c('rd', 'collate', 'namespace', 'vignette')))
         error <- try(devtools::build())
+        if(dir.exists(".git")) system("git --no-pager show -s --format='%h | %s | %an <%ae>' HEAD")
         if("try-error" %in% class(error)) {
-          message(".:: ",d," ",curversion," -> ",vkey$version," build failed ::.")
+          message(".:: ",fd," ",curversion," -> ",vkey$version," build failed ::.")
         } else {
-          message(".:: ",d," ",curversion," -> ",vkey$version," build success ::.")
+          message(".:: ",fd," ",curversion," -> ",vkey$version," build success ::.")
         }
-      } else message(".:: ",d," ",curversion," -> ",vkey$version," invalid commit ::.")
-    } else message(".:: ",d," ",curversion," ok ::.")
+      } else {
+        if(dir.exists(".git")) system("git --no-pager show -s --format='%h | %s | %an <%ae>' HEAD")
+        message(".:: ",fd," ",curversion," -> ",vkey$version," invalid commit ::.")
+      }
+    } else message(".:: ",fd," ",format(curversion, width=10)," ok ::.")
     setwd("..")
   }
   tools::write_PACKAGES()
